+++ libjcore/code/JWebBrowser.cpp
@@ -21,7 +21,7 @@
 //
 // "arrow --mailto \"$a\""
 
-#ifdef _J_OSX
+#ifdef _J_MACOS
 static const JUtf8Byte* kDefShowURLCmd = "open $u";
 #else
 static const JUtf8Byte* kDefShowURLCmd = "firefox $u";
@@ -29,7 +29,7 @@ static const JUtf8Byte* kDefShowURLCmd = "firefox $u";
 static const JUtf8Byte* kURLVarName    = "u";
 
 static const JUtf8Byte* kFileURLPrefix          = "file:";
-#ifdef _J_OSX
+#ifdef _J_MACOS
 static const JUtf8Byte* kDefShowFileContentCmd  = "open $f";
 #else
 static const JUtf8Byte* kDefShowFileContentCmd  = "firefox file:'$f'";
@@ -39,7 +39,7 @@ static const JUtf8Byte* kFileVarName            = "f";
 static const JUtf8Byte* kPathVarName            = "p";
 
 static const JUtf8Byte* kMailURLPrefix      = "mailto:";
-#ifdef _J_OSX
+#ifdef _J_MACOS
 static const JUtf8Byte* kDefComposeMailCmd  = "open mailto:$a";
 #else
 static const JUtf8Byte* kDefComposeMailCmd  = "firefox mailto:$a";