+++ libjx/code/JXLibVersion.h
@@ -63,6 +63,8 @@ static const char* kCurrentJXLibVersionStr = "4.0.0";
 //			urgent & idle tasks.  UI code must run only in the UI thread.
 //	JXImage
 //		Added CreateFromXPM() which works with > 256 colors.
+//	JXDisplay:
+//		*** Renamed IsOSX() to IsMacOS().
 
 // version 3.1.0:
 //	JXGetStringDialog: